Refactor post editing and creation modals; consolidate EditPostModal and CreatePostModal into PostFormModal. Update Home component to use the new unified modal. Enhance post handling in useHomePosts hook for better post creation and editing flow. Adjust API service methods to support form data for post creation and editing.
